refactor(stella-cli): split the wrapper report tests out of wrapper_plugin::tests
`wrapper_plugin/tests.rs` sits at the 1500-line ratchet, and this batch's two
new suites — the report lines a dispatch prints, and which process owes the
`--no-pipeline` notice — crossed it. They move to a submodule rather than a
baseline entry.

Also deletes three content-free constructions the prose guard named in prose
this batch touched.

Refs #3883
